requirements.txt
flask==1.1.2
numpy==1.20.3
Pillow==8.2.0
requests==2.25.1